+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.139 (2023-06-01)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.138 (2023-05-25)
+
+**Note:** Version bump only for package @parca/components
+
 ## 0.16.137 (2023-05-10)
 
 **Note:** Version bump only for package @parca/components